+++ src/schemathesis/service/__init__.py
@@ -16,3 +16,4 @@ from .constants import (
 from .events import Completed, Error, Event, Timeout
 from .handler import ServiceReporter
 from .models import TestRun
+from .report import ReportHandler